dev-tools/containers/data-upload/requirements.txt
crypt4gh==1.7
PyNaCl==1.6.1
python_lorem==1.3.0.post3
Requests==2.32.5
psycopg2-binary==2.9.10
